MovieRaterApi/requirements-dev.txt
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
config==0.5.1
dj-config-url==0.1.1
dj-static==0.0.6
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
future==0.18.2
mongoengine==0.24.1
pycparser==2.21
pymongo==3.12.3
python-decouple==3.6
pytz==2022.1
serializers==0.2.4
six==1.16.0
sqlparse==0.2.4
static3==0.7.0
view==0.1
